Fixed return types and added convenient method for lot quantity in ward

+++ src/main/java/org/isf/medicalstock/service/LotIoOperationRepository.java
@@ -40,10 +40,10 @@ public interface LotIoOperationRepository extends JpaRepository<Lot, String> {
 	Integer getMainStockQuantity(@Param("lot") Lot lot);
 	
 	@Query("select sum(w.in_quantity - w.out_quantity) FROM MedicalWard w WHERE w.id.lot = :lot")
-	Integer getWardsTotalQuantity(@Param("lot") Lot lot);
+	Double getWardsTotalQuantity(@Param("lot") Lot lot);
 	
 	@Query("select sum(w.in_quantity - w.out_quantity) FROM MedicalWard w WHERE w.id.lot = :lot and w.id.ward = :ward")
-	Integer getQuantityByWard(@Param("lot") Lot lot, @Param("ward") Ward ward);
+	Double getQuantityByWard(@Param("lot") Lot lot, @Param("ward") Ward ward);
 
 	@Query(value = "select LT_ID_A,LT_PREP_DATE,LT_DUE_DATE,LT_COST,"
 			+ "SUM(IF(MMVT_TYPE LIKE '%+%',MMV_QTY,-MMV_QTY)) as quantity from "
